hidden test corrections

+++ account/src/main/java/com/devsu/hackerearth/backend/account/controller/AccountController.java
@@ -40,7 +40,12 @@ public class AccountController {
 	public ResponseEntity<AccountDto> get(@PathVariable Long id){
 		// api/accounts/{id}
 		// Get accounts by id
-		return ResponseEntity.ok(accountService.getById(id));
+		AccountDto account = accountService.getById(id);
+		if (account == null){
+			return ResponseEntity.notFound().build();
+		}
+
+		return ResponseEntity.ok(account);
 	}
 
 	@PostMapping
@@ -55,10 +60,20 @@ public class AccountController {
 	public ResponseEntity<AccountDto> update(@PathVariable Long id, @RequestBody AccountDto accountDto){
 		// api/accounts/{id}
 		// Update accounts
-		accountDto.setId(id);
 
+		AccountDto accountExist = accountService.getById(id);
+		if (accountExist == null){
+			return ResponseEntity.notFound().build();
+		}
+		accountExist.setNumber(accountDto.getNumber());
+		accountExist.setType(accountDto.getType());
+		accountExist.setInitialAmount(accountDto.getInitialAmount());
+		accountExist.setActive(accountDto.isActive());
+		accountExist.setClientId(accountDto.getClientId());
+
+		AccountDto account = accountService.update(accountExist);
 		return ResponseEntity.ok(
-				accountService.update(accountDto)
+				account
 		);
 	}
 
@@ -66,16 +81,23 @@ public class AccountController {
 	public ResponseEntity<AccountDto> partialUpdate(@PathVariable Long id, @RequestBody PartialAccountDto partialAccountDto){
 		// api/accounts/{id}
 		// Partial update accounts
+		AccountDto updatedAccount = accountService.partialUpdate(id, partialAccountDto);
 
-		return ResponseEntity.ok(
-				accountService.partialUpdate(id, partialAccountDto)
-		);
+		if(updatedAccount == null){
+			return ResponseEntity.notFound().build();
+		}
+
+		return ResponseEntity.ok(updatedAccount);
 	}
 
 	@DeleteMapping("/{id}")
 	public ResponseEntity<Void> delete(@PathVariable Long id){
 		// api/accounts/{id}
 		// Delete accounts
+		AccountDto account = accountService.getById(id);
+		if (account == null){
+			return ResponseEntity.notFound().build();
+		}
 		accountService.deleteById(id);
 		return ResponseEntity.noContent().build();
 	}
